Move `Error` to its own module, impl Display/Error
I agree to license my contributions to each file under the terms given
at the top of each file I changed.
